Re-init frontend part of the project with Vite + React
