+++ frontend/src/App.tsx
@@ -1,0 +1,16 @@
+import './App.css'
+import UserAuthForm from './components/UserAuthform.tsx';
+
+function App() {
+
+  return (
+    <>
+      <h1>Who are you, warrior?</h1>
+      <div className="card">
+        <UserAuthForm/>
+      </div>
+    </>
+  )
+}
+
+export default App
