--- frontend/src/components/Header.tsx
@@ -1,4 +1,0 @@
-// components/Header.js
-export default function Header() {
-  return <h1>Welcome to the AlphaBeta!</h1>;
-}
